+++ packages/template/src/components-page/account-settings/email-and-auth/email-and-auth-page.tsx
@@ -1,7 +1,7 @@
+import { PageLayout } from "../page-layout";
 import { EmailsSection } from "./emails-section";
 import { MfaSection } from "./mfa-section";
 import { OtpSection } from "./otp-section";
-import { PageLayout } from "./page-layout";
 import { PasskeySection } from "./passkey-section";
 import { PasswordSection } from "./password-section";
