+++ packages/template/src/components-page/account-settings/email-and-auth/otp-section.tsx
@@ -1,8 +1,8 @@
 import { Button, Typography } from "@stackframe/stack-ui";
 import { useState } from "react";
-import { useStackApp, useUser } from "../../lib/hooks";
-import { useTranslation } from "../../lib/translations";
-import { Section } from "./section";
+import { useStackApp, useUser } from "../../../lib/hooks";
+import { useTranslation } from "../../../lib/translations";
+import { Section } from "../section";
 
 export function OtpSection() {
   const { t } = useTranslation();
