+++ packages/template/src/components-page/account-settings/settings/delete-account-section.tsx
@@ -1,8 +1,8 @@
 import { Accordion, AccordionContent, AccordionItem, AccordionTrigger, Button, Typography } from "@stackframe/stack-ui";
 import { useState } from "react";
-import { useStackApp, useUser } from "../../lib/hooks";
-import { useTranslation } from "../../lib/translations";
-import { Section } from "./section";
+import { useStackApp, useUser } from "../../../lib/hooks";
+import { useTranslation } from "../../../lib/translations";
+import { Section } from "../section";
 
 export function DeleteAccountSection() {
   const { t } = useTranslation();
