+++ packages/template/src/components-page/account-settings/settings/sign-out-section.tsx
@@ -1,7 +1,7 @@
 import { Button } from "@stackframe/stack-ui";
-import { useUser } from "../../lib/hooks";
-import { useTranslation } from "../../lib/translations";
-import { Section } from "./section";
+import { useUser } from "../../../lib/hooks";
+import { useTranslation } from "../../../lib/translations";
+import { Section } from "../section";
 
 export function SignOutSection() {
   const { t } = useTranslation();
