+++ packages/template/src/components-page/account-settings/teams/leave-team-section.tsx
@@ -1,9 +1,9 @@
 import { Button, Typography } from "@stackframe/stack-ui";
 import { useState } from "react";
-import { Team } from "../..";
-import { useUser } from "../../lib/hooks";
-import { useTranslation } from "../../lib/translations";
-import { Section } from "./section";
+import { Team } from "../../..";
+import { useUser } from "../../../lib/hooks";
+import { useTranslation } from "../../../lib/translations";
+import { Section } from "../section";
 
 export function LeaveTeamSection(props: { team: Team }) {
   const { t } = useTranslation();
